(feat): ensure organization separaring collections types
